Move createTeam and getTeams to admin use cases
