+++ src/use-cases/admin/index.ts
@@ -1,1 +1,3 @@
+export { createTeam, getTeams } from './teams'
+
 export { getUser, searchUsers } from './users'
