+++ src/use-cases/admin/teams.ts
@@ -1,0 +1,2 @@
+export { createTeam, getTeams } from '@/use-cases/user/teams'
+export type { CreateTeamParams } from '@/use-cases/user/teams'
