+++ src/use-cases/user/index.ts
@@ -5,8 +5,6 @@ export { getUser, updateUser } from './me'
 export { getTeamMember, getTeamMembers, updateTeamMember } from './members'
 
 export {
-  createTeam,
   getTeam,
-  getTeams,
   updateTeam,
 } from './teams'
